Fixed bug that was preventing sald from converging.
The problem was that, within sald, xtol was set to a fairly large value. It shouldn't be large,
because the state variable for population size is 1/2N. We need fairly tight tolerance on 1/2N to
get acceptable accuracy on 2N. With the old loose tolerance, sasimplex_converged would decide that
the simplex was small enough. However, there would still be large differences in function values,
so sasimplex_converged would return GSL_ETOLF--an indication that the simplex has collapsed to a
point but the function values still vary too much. This triggers a call to expand_around_best,
which increases the size of the simplex. This process continues until we reach the iteration
limit.

Bottom line: need to rethink convergence criteria. I think this is why the current code is slower
than the numerical recipes code.