Fixed the problem with sald stalling by increasing the lower bound on twoN and decreasing the upper bound on T. The problem arises when T/2N is large. These changes make sure we don't get trapped in that part of parameter space. This is a sensible thing to do, because the method is aimed at short time scales anyway, and it is reasonable to assume that 2N > 500.
For longer time scales, it might be useful to reduce the number of samples and
somehow scale things up. Tavare has a formula for the probability that there
are x lines of descent t generations ago. One could sum across that
distribution, calculating the ancient portion of the spectrum conditional on x.
You wouldn't need to consider large x, because those values would have low
probability.